+++ CHANGELOG.rst
@@ -3,7 +3,7 @@
 - implement Py3 compatibility
 - Allows tigger model lsm.html or text file to determine Dynamic Range
   using residual image
-- Compare input-out tigger models and plot correlation
+- Compare input-output flux tigger models and plot correlation
 
 0.1.1
 -----
@@ -12,6 +12,11 @@
 
 0.1.2
 -----
-- Use peak flux when computing DYnamic Range (DR)
+- Use peak flux when computing Dynamic Range (DR)
 - Compute three DR (local using deepest negative, local using rms, global uing rms)
 - Use a label instead of a path when creating stats dictionary
+
+0.2.0
+-----
+- Compare input-output astrometry tigger models and plot correlation
+- Compute correlation stats between output and input data-set